requirements.txt
numpy
# 2 times depracated
#pyuserinput 
pynput
pillow
